+++ docs/CHANGE_LOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.3.0:
 
+#### **Binding Options:**
+- Added a new binding option called "ignoreNullValues", which states if NULL values should be ignored and not rendered (defaults to false).
+- Added a new binding option called "ignoreFunctionValues", which states if FUNCTION values should be ignored and not rendered (defaults to false).
+
 #### **Public Functions:**
 - Added a new public function "refresh()", which refreshes the UI for a specific element.
 - Added a new public function "refreshAll()", which will refresh all the rendered elements.